Load data outside startServer module
It should not be the responsibility of the startServer function to also
know how to load the final data. It makes for cleaner code to plumb it
in from the outside so that the server related code can focus on just
launching the server and nothing more.
